requirements.txt
django==4.1.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
dj-database-url>=0.5,<0.6
gunicorn==20.1.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0